webapp/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
markdown==3.5.1
httpx==0.28.1
tiktoken==0.12.0
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.1.0
